Rename to separate functions from execution
