+++ src/fn/factory.js
